Implement Door abstract class
